requirements.txt
googletrans==3.0.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
python_pptx==0.6.23
Requests==2.31.0
six==1.16.0
